% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Main function}
\usage{
init()
}
\value{
The function returns an interactive shiny dashboard.
}
\description{
Main function
}
\details{
The function run the shiny music app including every features.
}
\author{
Edward Tandia
}
