% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Main function}
\usage{
main()
}
\value{
The function returns an interactive shiny dashboard.
}
\description{
Main function
}
\details{
The function run the shiny music app including every features.
}
\author{
Edward Tandia
}
